Added constructors

+++ src/app/Commande.java
@@ -6,7 +6,7 @@ import java.util.Date;
 public class Commande {
     protected String id;
     protected Date dateCreation;
-    protected int reduction;    // peut-etre changer le type?
+    protected double reduction;
     protected ArrayList<Emprunt> emprunts;
 
 }
